+++ RealPython/puppy_store/puppies/tests/test_views.py
@@ -1,0 +1,11 @@
+# holds all tests for our views
+import json
+from rest_framework import status
+from django.test import TestCase, Client
+from django.urls import reverse
+from ..models import Puppy
+from ..serializers import PuppySerializer
+
+
+# initialize the APIClient app
+client = Client()